isolate bash to Makefile, handle mocha PATH
Figure out where mocha is and assign it to variable.  Also, isolate bash
execution/interpretation to one a single environment.  There is some evidence
that the interpretation is slightly modified somehow; if scripts need to be
run, we can run them all from Makefile to that things can be tested locally by
hand as well as automatically.  Scripts are also ok, but this is a small enough
change the Makefile seems fine.